backend/requirements.txt
fastapi
uvicorn
requests
google-genai
nltk
python-dotenv
pydantic
praw
jugaad-data
